# internal/api/openapi.yaml
openapi: "3.0.3"
info:
  title: goscribe API
  version: "1.0.0"
  description: |
    HTTP API for goscribe — async audio transcription and AI post-processing.

    Submit a job with an audio file or existing transcript. The job is queued
    and processed asynchronously by a worker. Poll `GET /jobs/{id}` for results,
    or supply a `webhook_url` to receive them via HTTP POST on completion.
servers:
  - url: http://localhost:8080
    description: Local / Docker Compose
paths:
  /jobs:
    post:
      summary: Submit a transcription job
      description: |
        Accepts a multipart form with either an audio file to transcribe or
        an existing transcript to post-process. At least one of `file` or
        `transcript` is required.
      operationId: submitJob
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: |
                    Audio file to transcribe. Supported formats: mp3, mp4, mpeg,
                    mpga, m4a, wav, webm, ogg, flac, aac, aiff.
                    Maximum size is controlled by `MAX_UPLOAD_MB` (default 100 MB).
                transcript:
                  type: string
                  description: |
                    Existing transcript text to post-process directly.
                    Mutually exclusive with `file` — if both are provided,
                    `file` takes precedence and `transcript` is ignored.
                actions:
                  type: string
                  description: |
                    Comma-separated list of action IDs to run after transcription
                    (e.g. `openai-meeting-summary,openai-action-items`).
                    Pass `auto` to let the AI select the most relevant actions.
                    If omitted, only the raw transcript is produced.
                  example: openai-meeting-summary,openai-action-items
                provider:
                  type: string
                  enum: [openai, gemini]
                  description: |
                    AI provider to use for this job. Overrides the server default.
                    Automatic fallback to the other provider is enabled on the
                    worker side when both keys are configured.
                  example: openai
                webhook_url:
                  type: string
                  format: uri
                  description: |
                    URL to call via HTTP POST when the job completes.
                    The request body is the `JobResult` JSON object.
                    Requests to private/loopback addresses are rejected.
                  example: https://example.com/hooks/transcription
      responses:
        "202":
          description: Job accepted and queued for processing.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitResponse"
              example:
                job_id: "3f2504e0-4f89-11d3-9a0c-0305e82c3301"
                status: queued
        "400":
          $ref: "#/components/responses/BadRequest"
        "413":
          $ref: "#/components/responses/TooLarge"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  schemas:
    SubmitResponse:
      type: object
      required: [job_id, status]
      properties:
        job_id:
          type: string
          format: uuid
          description: Unique identifier for the submitted job.
        status:
          type: string
          enum: [queued]
          description: Always `queued` on a successful submission.
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message.
  responses:
    BadRequest:
      description: Invalid request — missing required fields or malformed input.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "provide either a file or transcript text"
...
    TooLarge:
      description: Uploaded file exceeds the configured size limit.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "file exceeds 100 MB limit"
    InternalError:
      description: Unexpected server-side error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "enqueuing job: connection refused"